split qr demo into screen and widget
